Move initialize jwt to jwt directory
